moved start conversation to the modals folder
