+++ packages/client/src/app/components/modals/portfolio/start-conversation/start-conversation.component.ts
@@ -3,8 +3,8 @@ import { FormGroup, FormControl, Validators } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 import { CreateInitialMessage } from '@pulp-fiction/models/messages';
-import { MessagesService } from '../../../services/content';
-import { AlertsService } from '../../../modules/alerts';
+import { MessagesService } from '../../../../services/content';
+import { AlertsService } from '../../../../modules/alerts';
 
 @Component({
   selector: 'pulp-fiction-start-conversation',
